+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/time/SystemTimeProvider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.utilities.time;
+package co.elastic.apm.android.sdk.internal.time;
 
 public class SystemTimeProvider {
