+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/time/ntp/TrueTimeBootBroadcastReceiver.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.utilities.time.ntp;
+package co.elastic.apm.android.sdk.internal.time.ntp;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
