+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/session/impl/DefaultSessionIdProvider.java
@@ -28,7 +28,7 @@ import co.elastic.apm.android.sdk.internal.api.Initializable;
 import co.elastic.apm.android.sdk.internal.utilities.providers.Provider;
 import co.elastic.apm.android.sdk.internal.services.Service;
 import co.elastic.apm.android.sdk.internal.services.preferences.PreferencesService;
-import co.elastic.apm.android.sdk.internal.utilities.time.SystemTimeProvider;
+import co.elastic.apm.android.sdk.internal.time.SystemTimeProvider;
 import co.elastic.apm.android.sdk.session.SessionIdProvider;
 
 /**
